feat(daemon): implement pool link counting
This implements link counting for the pool. This is needed to track when package should be removed from the pool.

The counter increases each time a package that references the pool entry is added to the repository and decreases when a package is removed. When the counter reaches zero, the package is removed from the pool.