ci: run clang-tidy
We run clang-tidy on every change. For that we introduce a separate CMake
preset that is selected in the CI and activates the clang-tidy support in
CMake. Additionally a Python script is provided to run clang-tidy locally
when the build already exists.

For now we enforce only the warning of one check as an error, but we aim to
increase that over time.